tidied up ramp_utils
always good to banish anything with 'utils' in the name!
